+++ src/client/index.ts
@@ -10,14 +10,14 @@ const socket = io();
 // Create view
 const view = new View();
 
-const game = new Game(view,  DEBUG_MODE);
+let gotWorld = false;
+console.log("Waiting for world");
 
-/*console.log("Waiting for world");
-socket.on("world", cluster => {
+socket.on("world", clusterString => {
+    if (gotWorld) return;
     console.log("Got world");
+    gotWorld = true;
     
     // Create game server
-    const game = new Game(view, cluster, DEBUG_MODE);
+    const game = new Game(view, clusterString, DEBUG_MODE);
 });
-
-*/