+++ src/share/data/coordinate/ICoordinate.ts
@@ -1,0 +1,8 @@
+/**
+ * The base interface for any type of immutable coordinate.
+ */
+export interface ICoordinate {
+  readonly x: number;
+  readonly y: number;
+  readonly z: number;
+}