+++ src/share/data/grid/chunk-grid/index.ts
@@ -1,2 +1,0 @@
-export { ChunkGrid } from "./ChunkGrid";
-export { IChunkGrid } from "./IChunkGrid";