+++ src/share/data/grid/index.ts
@@ -1,3 +1,2 @@
 export { Grid } from "./Grid";
 export { IGrid } from "./IGrid";
-export * from "./absolute-grid";