+++ src/share/data/grid/relative-grid/index.ts
@@ -1,2 +1,0 @@
-export { RelativeGrid } from "./RelativeGrid";
-export { IRelativeGrid } from "./IRelativeGrid";