+++ changelog.md
@@ -1,15 +1,24 @@
 ## Backlog
-* Send world data between client and server
 * Sound FX
 * Fix player movement behavior (going under arches, placing blocks under self, etc.)
-* Make asset manager static
-* Remove coordinate system and greatly simplify grid system
+* Remove different types of coordinates and add converter class
+
+## 0.1.1
+
+* TODO: Make asset manager static
+* TODO: Changes in blocks are automatically synched between server and all clients
+
+## 0.1.0
+
+* Re-write grid system using generic parameters
+* Serverside block storage
+* Server sends block data to clients on logging in
+* Shadows
 
 ## 0.0.4
 
 * Fixed lines between blocks
 * Improved top-down biome generation (inverse instead of logistic)
-
 
 ## 0.0.3